Better touch support and Drag func update
Functions added and used to register both touch and mouse events for elements like buttons.

Drag function was also updated to not explicitly work off of event target. A list of handles can now be defined that move and element. (bubbling not allowed. all handles must be specified even if child of element)