drag function and card image update
Drag function updated to allow for more control over handles by passing in an object with a handle and a "transparent" property. This determines if children block drag actions.

Image styling within project cards updated to include a wrapper for scaling image based on height. This is for easier changes if card height changes.